requirements.txt
# ==========================================
# SmartDoc RAG — Python Dependencies
# ==========================================

# --- API Framework ---
fastapi>=0.135.3
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# --- LangChain Ecosystem ---
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.8
langchain-text-splitters>=0.2.0
langgraph>=0.1.9

# --- Vector Store ---
faiss-cpu>=1.8.0

# --- Embeddings & NLP ---
sentence-transformers>=3.0.0

# --- Document Processing ---
pymupdf>=1.24.0
pillow>=10.3.0

# --- OCR ---
easyocr>=1.7.1

# --- Web Search ---
tavily-python>=0.3.3
duckduckgo-search>=6.1.0

# --- Utilities ---
python-dotenv>=1.0.1
pydantic>=2.7.0
httpx>=0.27.0
numpy>=1.26.0